Added story "Lore" folder on key "story_lore_folder" of "story_info" dictionary in "Text File Reader.php" line 444. Changed these variables to keys on the "story_info" dictionary: author_name readers readers_number chapter_number (changed "chapters" variable to the value of "chapter_number" key) author_number_text chapter_number_text readers_number_text
Added a new key for the full default author of the story on the "person_names" array.

Created four new arrays on "Settings Checker.php":
website_setting_names (array with setting names)
portuguese_website_setting_names (keyed array with translated setting names)
story_website_setting_names (array with setting names)
portuguese_story_website_setting_names (keyed array with translated setting names)

Created a new function on "Settings Checker.php" called "Check_Website_Setting".
This is to check the website setting on _POST only after the "Website.php" file has been loaded (as the file often changes website settings).
Changed how the website settings are checked on _POST.
Before, they would be checked by a keyed array, but you can not create keyed arrays with HTML input.
So the array on _POST only supports one setting value to be changed, it will change the setting to the opposite of its current value.
If the setting value is False, it will change it to True.
Added "website_setting" and "story_website_setting" select elements on the form on "Select Website Form.php" file.
The default website_setting is "notifications" (deactivates notifications), and the default story_website_setting is "nothing".

Created a way for the "Website.php" file to define its theme-color for the meta tag.
Defining it on the "theme_color" key of the "website_info" keyed array.
The "V_Global.php" file checks if the key "theme_color" is set.
If it is not, then it sets it as "#916f3b" (color for "The Life of Littletato" website).
Created a way to hide Read and Comment Modals by pressing the Escape key, on the "Hide Modal.js" file on story website JavaScript files folder.
Edited the "Hide Chapter Button Images.js" file and renamed it to "Hide Chapter Covers.js" in the root folder of the Stake2 website.
Moved it to the "Story Websites" JavaScript folder, and moved it to the Story Website section of JavaScript links.
Created the button on top of Chapter Button Tabs on story websites to hide chapter covers on the Chapter Buttons Tab.
Added a new story website setting called "show_story_covers_on_chapter_buttons_tab" on "story_website_settings" keyed array.
Made its default value "False", to hide chapter covers on the Chapter Buttons Tab, making the website lighter.
Changed the line number of the first watched video in 2022 to 13 on the "Watch History" website.