I made the Comment Template better and made it possible to use it on the normal website comments. Moved Comment Template to "Forms.php". Created a way for "Settings Checker" to dynamically create the Portuguese setting name and setting value arrays, using the English setting names array. Removed duplicate code in "Comments Generator", adding only one while block and an if to check if there is only one chapter comment. Also removed duplicate code in "Chapter Comment And Read Displayer", adding an if in the same way as above. Changed the way that dates of website comments and website dates are displayed, with a "Language_Item_Definer" variable with the format of the language. The month before day on English websites, and the day before the month on Portuguese websites. Improved the style of chapter reading and comment forms, and made the code of the modals better. Each chapter modal is now generated right after the chapter tab, and not after all chapter tabs have been generated.
Renamed "Chapters Generator" to "Tab Generator", and "Tab Generator" to "Tab Content Generator".
All of these are story website files.

Moved these PHP files to the "Chapter Files" folder of the "Story Website Files" folder and removed the "Chapter" part of the filenames:
"Button Generator", "Open Chapter Script", "Tab Generator", "Text Displayer"

Changed the way the story website Tab Generator generates chapter tabs.
Removed a lot of garbage code and made it more dynamic and beautiful.

Renamed these folders on the "Story Website Files" folder:
"Reader And Reads Files" to "Reading Folder"
"Comments Folder" to "Comment Folder"

Remade the "Redirect To Website Of User Language.js" JavaScript file.
It is on the root website folder in the "JavaScript/Function" folder.
Renamed it to "Language Redirector.js".
Coded it using an array of English website names and a property (dictionary) of Portuguese website names with English website names as keys.
Used the Fetch API to read the two files in the root website folder in the "Texts/Websites List" folder.
And JSON.parse to parse the files into array and property (dictionary).
Now every website needs to be only inside of the English website folder for the Redirect to work.
The "European Portuguese" (PT-PT) language will no more use a different URL and folder name, it will be the same as the "PT-BR" language.
Created two new functions, "English_Language_Check", and "Portuguese_Language_Check".
To check if the user language is English or Portuguese.
If the website is the index one, redirect to the user language website.
If the website title is in English and the user language is Portuguese, redirect to the Portuguese website.
If the website title is in Portuguese and the user language is English, redirect to the English website.
Created a new function called "Redirect" to redirect to the user language website and also add the URL addon.

Created a way to hide the big space in computer mode in the "Show Hide.js" JavaScript file on the root website folder in the "JavaScript/Function" folder.
By creating a div with br elements with the id "computer_buttons_bar_space" in "Body.html" RainTPL template.
Hiding it when clicking the hide button, showing it when clicking on the show button, on the JavaScript hide and show functions, respectively.